+++ packages/common/src/schemas/index.ts
@@ -7,8 +7,8 @@ export { getGhanaDocuments } from './documents/workflow/documents/schemas/GH';
 export {
   findDocumentSchemaByTypeAndCategory,
   getDocumentId,
+  getDocumentSchemaByCountry,
   getDocumentsByCountry,
 } from './documents/workflow/documents/schemas/index';
-export { type TDocument } from './documents/workflow/documents/types';
-export { type TAvailableDocuments } from './documents/workflow/documents/types';
-export { getDocumentSchemaByCountry } from './documents/workflow/documents/schemas/index';
+export { type TAvailableDocuments, type TDocument } from './documents/workflow/documents/types';
+export { WorkflowDefinitionConfigThemeSchema } from './workflow/workflow-config-theme';