+++ services/workflows-service/src/alert/alert.service.intg.test.ts
@@ -5,6 +5,7 @@ import {
   Customer,
   PaymentMethod,
   Project,
+  TransactionDirection,
   TransactionRecordType,
 } from '@prisma/client';
 import { tearDownDatabase } from '@/test/helpers/database-helper';
@@ -40,15 +41,12 @@ describe('AlertService', () => {
         AlertRepository,
         AlertDefinitionRepository,
         AlertService,
-        TransactionFactory,
       ],
     }).compile();
 
     prismaService = module.get(PrismaService);
 
     alertService = module.get(AlertService);
-
-    transactionFactory = module.get(TransactionFactory);
   });
 
   beforeEach(async () => {
@@ -66,6 +64,11 @@ describe('AlertService', () => {
     );
 
     project = await createProject(prismaService, customer, faker.datatype.uuid());
+
+    transactionFactory = new TransactionFactory({
+      prisma: prismaService,
+      projectId: project.id,
+    });
   });
 
   afterAll(tearDownDatabase);
@@ -75,10 +78,158 @@ describe('AlertService', () => {
 
     beforeEach(() => {
       baseTransactionFactory = transactionFactory
-        .project(project)
         .paymentMethod(PaymentMethod.credit_card)
-        .withEndUserBeneficiary()
         .transactionDate(faker.date.recent(6));
+    });
+
+    describe('Rule: STRUC_CC', () => {
+      let alertDefinition: AlertDefinition;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              label: 'STRUC_CC',
+              ...ALERT_DEFINITIONS.STRUC_CC,
+            },
+            project,
+          ),
+        });
+      });
+
+      test('When there are more than 5 inbound transactions with amount of 500, an alert should be created', async () => {
+        // Arrange
+        const transactions = await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(500)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .count(6)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(transactions[0]?.counterpartyBeneficiaryId);
+      });
+
+      test('When there are less than 5 inbound transactions with amount of 500, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(500)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .count(4)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+
+      test('When there are more than 5 inbound transactions with amount of 499, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(499)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .count(6)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+    });
+
+    describe('Rule: STRUC_APM', () => {
+      let alertDefinition: AlertDefinition;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              label: 'STRUC_APM',
+              ...ALERT_DEFINITIONS.STRUC_APM,
+            },
+            project,
+          ),
+        });
+      });
+
+      test('When there are more than 5 inbound transactions with amount of 500, an alert should be created', async () => {
+        // Arrange
+        const transactions = await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(500)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.pay_pal)
+          .count(6)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(transactions[0]?.counterpartyBeneficiaryId);
+      });
+
+      test('When there are less than 5 inbound transactions with amount of 500, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(500)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.pay_pal)
+          .count(4)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+
+      test('When there are more than 5 inbound transactions with amount of 499, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .withBusinessBeneficiary()
+          .withEndUserOriginator()
+          .amount(499)
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.pay_pal)
+          .count(6)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
     });
 
     describe('Rule: CHVC_C', () => {
@@ -94,16 +245,16 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(15)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -121,16 +272,16 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -154,18 +305,18 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .amount(100)
           .count(51)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .amount(100)
           .count(49)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -183,18 +334,18 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .amount(100)
           .count(49)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .amount(100)
           .count(49)
-          .create({
-            transactionType: TransactionRecordType.chargeback,
-          });
+          .type(TransactionRecordType.chargeback)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -218,16 +369,16 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(15)
-          .create({
-            transactionType: TransactionRecordType.refund,
-          });
+          .type(TransactionRecordType.refund)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.refund,
-          });
+          .type(TransactionRecordType.refund)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -245,16 +396,16 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.refund,
-          });
+          .type(TransactionRecordType.refund)
+          .create();
         const business2Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .count(14)
-          .create({
-            transactionType: TransactionRecordType.refund,
-          });
+          .type(TransactionRecordType.refund)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -278,19 +429,27 @@ describe('AlertService', () => {
         // Arrange
         const business1Transactions = await baseTransactionFactory
           .withBusinessOriginator()
+          .withEndUserBeneficiary()
           .amount(1000)
           .count(11)
-          .create({
-            transactionType: TransactionRecordType.refund,
-          });
+          .type(TransactionRecordType.refund)
+          .create();
 
-        await baseTransactionFactory.withBusinessOriginator().amount(10).count(12).create({
-          transactionType: TransactionRecordType.refund,
-        });
+        await baseTransactionFactory
+          .withBusinessOriginator()
+          .withEndUserBeneficiary()
+          .amount(10)
+          .count(12)
+          .type(TransactionRecordType.refund)
+          .create();
 
-        await baseTransactionFactory.withBusinessOriginator().amount(5001).count(13).create({
-          transactionType: TransactionRecordType.chargeback,
-        });
+        await baseTransactionFactory
+          .withBusinessOriginator()
+          .withEndUserBeneficiary()
+          .amount(5001)
+          .count(13)
+          .type(TransactionRecordType.chargeback)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -315,9 +474,13 @@ describe('AlertService', () => {
 
       it('When the sum of refunds amount is less than 5000, no alert should be created', async () => {
         // Arrange
-        await baseTransactionFactory.withBusinessOriginator().amount(100).count(49).create({
-          transactionType: TransactionRecordType.refund,
-        });
+        await baseTransactionFactory
+          .withBusinessOriginator()
+          .withEndUserBeneficiary()
+          .amount(100)
+          .count(49)
+          .type(TransactionRecordType.refund)
+          .create();
 
         // Act
         await alertService.checkAllAlerts();
@@ -361,15 +524,17 @@ describe('AlertService', () => {
 
       it('When there are >=3 chargeback transactions and they are >=50% of the total transactions, an alert should be created', async () => {
         // Arrange
-
         const chargebackTransactions = await baseTransactionFactory
           .type(TransactionRecordType.chargeback)
           .withCounterpartyOriginator(counteryparty.id)
+          .withEndUserBeneficiary()
           .count(3)
           .create();
+
         await baseTransactionFactory
           .type(TransactionRecordType.payment)
           .withCounterpartyOriginator(counteryparty.id)
+          .withEndUserBeneficiary()
           .count(3)
           .create();
 
@@ -389,11 +554,14 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .type(TransactionRecordType.chargeback)
+          .withEndUserBeneficiary()
           .withCounterpartyOriginator(counteryparty.id)
           .count(3)
           .create();
+
         await baseTransactionFactory
           .type(TransactionRecordType.payment)
+          .withEndUserBeneficiary()
           .withCounterpartyOriginator(counteryparty.id)
           .count(4)
           .create();
@@ -410,13 +578,194 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .type(TransactionRecordType.chargeback)
+          .withEndUserBeneficiary()
           .withCounterpartyOriginator(counteryparty.id)
           .count(2)
           .create();
+
         await baseTransactionFactory
           .type(TransactionRecordType.payment)
+          .withEndUserBeneficiary()
           .withCounterpartyOriginator(counteryparty.id)
           .count(2)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+    });
+
+    describe('Rule: TLHAICC', () => {
+      let alertDefinition: AlertDefinition;
+      let counteryparty: Counterparty;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              label: 'TLHAICC',
+              ...ALERT_DEFINITIONS.TLHAICC,
+            },
+            project,
+          ),
+        });
+
+        const correlationId = faker.datatype.uuid();
+        counteryparty = await prismaService.counterparty.create({
+          data: {
+            project: { connect: { id: project.id } },
+            correlationId: correlationId,
+            business: {
+              create: {
+                correlationId: correlationId,
+                companyName: faker.company.name(),
+                registrationNumber: faker.datatype.uuid(),
+                mccCode: faker.datatype.number({ min: 1000, max: 9999 }),
+                businessType: faker.lorem.word(),
+                project: { connect: { id: project.id } },
+              },
+            },
+          },
+        });
+      });
+
+      it('When there are >2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(2)
+          .create();
+
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(counteryparty.id);
+      });
+
+      it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(1)
+          .create();
+
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+    });
+
+    describe('Rule: TLHAIAPM', () => {
+      let alertDefinition: AlertDefinition;
+      let counteryparty: Counterparty;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              label: 'TLHAIAPM',
+              ...ALERT_DEFINITIONS.TLHAIAPM,
+            },
+            project,
+          ),
+        });
+
+        const correlationId = faker.datatype.uuid();
+        counteryparty = await prismaService.counterparty.create({
+          data: {
+            project: { connect: { id: project.id } },
+            correlationId: correlationId,
+            business: {
+              create: {
+                correlationId: correlationId,
+                companyName: faker.company.name(),
+                registrationNumber: faker.datatype.uuid(),
+                mccCode: faker.datatype.number({ min: 1000, max: 9999 }),
+                businessType: faker.lorem.word(),
+                project: { connect: { id: project.id } },
+              },
+            },
+          },
+        });
+      });
+
+      it('When there are >2 APM transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.apple_pay)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(2)
+          .create();
+
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.pay_pal)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(counteryparty.id);
+      });
+
+      it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.google_pay)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(1)
+          .create();
+
+        await baseTransactionFactory
+          .direction('inbound')
+          .paymentMethod(PaymentMethod.bank_transfer)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
           .create();
 
         // Act