+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ballerine/backoffice-v2
 
+## 0.7.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.4
+  - @ballerine/workflow-browser-sdk@0.6.7
+  - @ballerine/workflow-node-sdk@0.6.7
+
+## 0.7.5
+
+### Patch Changes
+
+- Added workflow definition theme schemas
+- Updated dependencies
+  - @ballerine/common@0.9.3
+  - @ballerine/workflow-browser-sdk@0.6.6
+  - @ballerine/workflow-node-sdk@0.6.6
+
 ## 0.7.4
 
 ### Patch Changes